+++ src/spyglass/spikesorting/v0/spikesorting_recording.py
@@ -1,4 +1,3 @@
-from functools import reduce
 from pathlib import Path
 from shutil import rmtree as shutil_rmtree
 
@@ -12,12 +11,7 @@ from tqdm import tqdm
 
 from spyglass.common.common_device import Probe, ProbeType  # noqa: F401
 from spyglass.common.common_ephys import Electrode, ElectrodeGroup
-from spyglass.common.common_interval import (
-    IntervalList,
-    interval_list_intersect,
-    intervals_by_length,
-    union_adjacent_index,
-)
+from spyglass.common.common_interval import Interval, IntervalList
 from spyglass.common.common_lab import LabTeam  # noqa: F401
 from spyglass.common.common_nwbfile import Nwbfile
 from spyglass.common.common_session import Session  # noqa: F401
@@ -247,6 +241,12 @@ class SortInterval(SpyglassMixin, dj.Manual):
     sort_interval: longblob # 1D numpy array with start and end time for a single interval to be used for spike sorting
     """
 
+    def fetch_interval(self):
+        """Fetch interval list object for a given key."""
+        if not len(self) == 1:
+            raise ValueError(f"Expected one row, got {len(self)}")
+        return Interval(self.fetch1("sort_interval"))
+
     # NOTE: See #630, #664. Excessive key length.
 
 
@@ -337,15 +337,13 @@ class SpikeSortingRecording(SpyglassMixin, dj.Computed):
         """
         rec_info = self._make_file(key)
 
-        IntervalList.insert1(
-            {
-                "nwb_file_name": key["nwb_file_name"],
-                "interval_list_name": rec_info["name"],
-                "valid_times": self._get_sort_interval_valid_times(key),
-                "pipeline": "spikesorting_recording_v0",
-            },
-            replace=True,
+        sort_interval_valid_times = self._get_sort_interval_valid_times(key)
+        sort_interval_valid_times.set_key(
+            nwb_file_name=key["nwb_file_name"],
+            interval_list_name=rec_info["name"],
+            pipeline="spikesorting_recording_v0",
         )
+        IntervalList.insert1(sort_interval_valid_times.as_dict, replace=True)
 
         self_insert = dict(
             key,
@@ -498,7 +496,7 @@ class SpikeSortingRecording(SpyglassMixin, dj.Computed):
                 "nwb_file_name": nwb_file_name,
                 "sort_interval_name": sort_interval_name,
             }
-        ).fetch1("sort_interval")
+        ).fetch_interval()
 
         valid_interval_times = (
             IntervalList
@@ -506,16 +504,14 @@ class SpikeSortingRecording(SpyglassMixin, dj.Computed):
                 "nwb_file_name": key["nwb_file_name"],
                 "interval_list_name": interval_list_name,
             }
-        ).fetch1("valid_times")
+        ).fetch_interval()
 
-        valid_sort_times = interval_list_intersect(
-            sort_interval, valid_interval_times
-        )
+        valid_sort_times = sort_interval.intersect(valid_interval_times)
+
         # Exclude intervals shorter than specified length
-        if "min_segment_length" in params:
-            valid_sort_times = intervals_by_length(
-                valid_sort_times, min_length=params["min_segment_length"]
-            )
+        if min_length := params.get("min_segment_length"):
+            valid_sort_times = valid_sort_times.by_length(min_length=min_length)
+
         return valid_sort_times
 
     def _get_filtered_recording(self, key: dict):
@@ -540,7 +536,7 @@ class SpikeSortingRecording(SpyglassMixin, dj.Computed):
             nwb_file_abs_path, load_time_vector=True
         )
 
-        valid_sort_times = self._get_sort_interval_valid_times(key)
+        valid_sort_times = self._get_sort_interval_valid_times(key).times
         # shape is (N, 2)
         valid_sort_times_indices = np.array(
             [
@@ -549,13 +545,11 @@ class SpikeSortingRecording(SpyglassMixin, dj.Computed):
             ]
         )
         # join intervals of indices that are adjacent
-        valid_sort_times_indices = reduce(
-            union_adjacent_index, valid_sort_times_indices
+        valid_sort_times_indices = (
+            Interval(valid_sort_times_indices)
+            .union_adjacent_consolidate()
+            .times
         )
-        if valid_sort_times_indices.ndim == 1:
-            valid_sort_times_indices = np.expand_dims(
-                valid_sort_times_indices, 0
-            )
 
         # create an AppendRecording if there is more than one disjoint sort interval
         if len(valid_sort_times_indices) > 1: